Added main recon daemon code.
Peer is the class which identify the hosts as peer and performs
the operations on the dataset.

Connection Manager is the singleton administering the connection and
effectively sending and receiving data.

Message is a class internal to the connection manager which encloses
the data to be send.

Other structures are helper such as Communication and some message
type.