[IMP] Key removal and blocklist
This commit add support for key removal based on the web interface via
challenge for users or command line for admin

Users will be able to submit a clearsign PGP message along with the
keyid they wish to remove. Once the signature has been verified the
system will proceed to wipe out the key material from the database.
All data related to the key will be wiped out as well.
This required some tuning in the database creation schema to support
cascade delete of records

Moreover, the removed key will be added into a blocklist, stored in
database to prevent adding the key again, both from web interface or
reconciliation.

Admin, on the backend side, will make use of the `blocklist` command to
add a list of key IDs to block. The system will remove the matching keys
as well

Fixing also the transaction mechanism that disabled unique checks
allowing duplication of signatures (wasted space)